% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported-data.R
\docType{data}
\name{BENCH.REGIONS}
\alias{BENCH.REGIONS}
\title{Dataset to bridge from country to all existing bench regions it is in}
\format{An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1295 rows and 4 columns.}
\source{
https://github.com/2DegreesInvesting/Reference/blob/master/ReferenceData/BenchRegions.csv
}
\usage{
BENCH.REGIONS
}
\description{
Dataset to bridge from country to all existing bench regions it is in
}
\examples{
library(dplyr)

BENCH.REGIONS

# The `CountryISO` of Namibia is the literal string "NA"
BENCH.REGIONS \%>\%
  filter(Country == "Namibia")
}
\seealso{
Other exported datasets: \code{\link{INDEX.REGIONS}}

Other possible_snapshots: \code{\link{ALD.BV}},
  \code{\link{ALD.CB}}, \code{\link{ALD.CC}},
  \code{\link{ALD.Company}}, \code{\link{ALD.EQ}},
  \code{\link{ALD.SPV}}, \code{\link{BALANCE.SHEET.DATA}},
  \code{\link{BicsSectorBridge}}, \code{\link{CB_OG}},
  \code{\link{DebtMarketClimate}},
  \code{\link{DebtMarket}}, \code{\link{EQMarket.Size}},
  \code{\link{EQ_OG}}, \code{\link{FIN.DATA}},
  \code{\link{Fund.Data}}, \code{\link{FundsTrusts}},
  \code{\link{GROUPS.GOVT}}, \code{\link{INDEX.REGIONS}},
  \code{\link{Indices}}, \code{\link{LoanMarketClimate}},
  \code{\link{LoanMarket}}, \code{\link{PHYSICAL.RISK.CB}},
  \code{\link{PHYSICAL.RISK.EQ}}, \code{\link{Receipts}},
  \code{\link{RevenueSplit}}, \code{\link{SCENLong}},
  \code{\link{SCEN}}, \code{\link{SEC.TYPE.BONDS}},
  \code{\link{SectorBridge}}, \code{\link{TYPE.BONDS}},
  \code{\link{TYPE.EQUITY}}, \code{\link{TYPE.OTHERS}},
  \code{\link{TYPE.RECEIPTS}}, \code{\link{sector.bridge}}
}
\concept{exported datasets}
\concept{possible_snapshots}
\keyword{datasets}
