% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_private_site.R
\name{browse_private_site}
\alias{browse_private_site}
\title{Browse private site}
\usage{
browse_private_site()
}
\value{
Invisible.
}
\description{
Opens a web browser on the home page of this package's website.
}
\examples{
# Use interactively only
# browse_private_site()
}
\seealso{
Other miscellaneous utility functions: \code{\link{build_private_site}},
  \code{\link{degrees}}, \code{\link{set_version}}
}
\concept{miscellaneous utility functions}
