% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_private_site.R
\name{build_private_site}
\alias{build_private_site}
\title{Build private website}
\usage{
build_private_site(run_dont_run = TRUE)
}
\arguments{
\item{run_dont_run}{Run examples that are surrounded in \\dontrun?}
}
\value{
Nothing. It's called for its side effect.
}
\description{
Builds docs/ into inst/ so it's accessible locally.
}
\examples{
# Use interactively only
# build_private_site()
}
\seealso{
Other miscellaneous utility functions: \code{\link{browse_private_site}},
  \code{\link{degrees}}, \code{\link{set_version}}
}
\concept{miscellaneous utility functions}
