% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionary.R
\docType{data}
\name{data_dictionary}
\alias{data_dictionary}
\title{Data dictionary}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17 rows and 3 columns.}
\usage{
data_dictionary
}
\description{
This dataset provides metadata to understand other datasets.
}
\examples{
# Metadata for `data_dictionary` itself
data_dictionary \%>\%
  dplyr::filter(dataset == "data_dictionary")

data_dictionary
}
\keyword{datasets}
