% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_loanbook.R
\docType{data}
\name{fake_loanbook}
\alias{fake_loanbook}
\title{A fake loanbook dataset}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 320 rows and 14 columns.}
\usage{
fake_loanbook
}
\description{
A dataframe containing fake loanbook data. This is a public dataset useful
for examples and tests.
}
\examples{
fake_loanbook
}
\seealso{
data_dictionary

Other bank datasets: \code{\link{loanbook_description}}
}
\concept{bank datasets}
\keyword{datasets}
