% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-large.R
\name{Fund.Data}
\alias{Fund.Data}
\title{Dataset}
\usage{
Fund.Data()
}
\description{
Dataset
}
\examples{
\dontrun{
Fund.Data()
}
}
\seealso{
Other datasets in directories other than Reference or datastore: \code{\link{DebtMarketClimate}},
  \code{\link{DebtMarket}}, \code{\link{EQMarket.Size}},
  \code{\link{Indices}}, \code{\link{LoanMarketClimate}},
  \code{\link{LoanMarket}}, \code{\link{PHYSICAL.RISK.CB}},
  \code{\link{PHYSICAL.RISK.EQ}}, \code{\link{SCENLong}},
  \code{\link{SCEN}}

Other possible_snapshots: \code{\link{ALD.BV}},
  \code{\link{ALD.CB}}, \code{\link{ALD.CC}},
  \code{\link{ALD.Company}}, \code{\link{ALD.EQ}},
  \code{\link{ALD.SPV}}, \code{\link{BALANCE.SHEET.DATA}},
  \code{\link{BENCH.REGIONS}},
  \code{\link{BicsSectorBridge}}, \code{\link{CB_OG}},
  \code{\link{DebtMarketClimate}},
  \code{\link{DebtMarket}}, \code{\link{EQMarket.Size}},
  \code{\link{EQ_OG}}, \code{\link{FIN.DATA}},
  \code{\link{FundsTrusts}}, \code{\link{GROUPS.GOVT}},
  \code{\link{INDEX.REGIONS}}, \code{\link{Indices}},
  \code{\link{LoanMarketClimate}},
  \code{\link{LoanMarket}}, \code{\link{PHYSICAL.RISK.CB}},
  \code{\link{PHYSICAL.RISK.EQ}}, \code{\link{Receipts}},
  \code{\link{RevenueSplit}}, \code{\link{SCENLong}},
  \code{\link{SCEN}}, \code{\link{SEC.TYPE.BONDS}},
  \code{\link{SectorBridge}}, \code{\link{TYPE.BONDS}},
  \code{\link{TYPE.EQUITY}}, \code{\link{TYPE.OTHERS}},
  \code{\link{TYPE.RECEIPTS}}, \code{\link{sector.bridge}}
}
\concept{datasets in directories other than Reference or datastore}
\concept{possible_snapshots}
