% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported-data.R
\docType{data}
\name{INDEX.REGIONS}
\alias{INDEX.REGIONS}
\title{Dataset to identify the equity index regions based on the MSCI factsheets}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 327 rows and 3 columns.}
\source{
https://github.com/2DegreesInvesting/Reference/blob/master/ReferenceData/IndexRegions.rda
}
\usage{
INDEX.REGIONS
}
\description{
Dataset to identify the equity index regions based on the MSCI factsheets,
e.g. 23 countries within the MSCI World.
}
\examples{
INDEX.REGIONS
}
\seealso{
Other exported datasets: \code{\link{BENCH.REGIONS}}

Other possible_snapshots: \code{\link{ALD.BV}},
  \code{\link{ALD.CB}}, \code{\link{ALD.CC}},
  \code{\link{ALD.Company}}, \code{\link{ALD.EQ}},
  \code{\link{ALD.SPV}}, \code{\link{BALANCE.SHEET.DATA}},
  \code{\link{BENCH.REGIONS}},
  \code{\link{BicsSectorBridge}}, \code{\link{CB_OG}},
  \code{\link{DebtMarketClimate}},
  \code{\link{DebtMarket}}, \code{\link{EQMarket.Size}},
  \code{\link{EQ_OG}}, \code{\link{FIN.DATA}},
  \code{\link{Fund.Data}}, \code{\link{FundsTrusts}},
  \code{\link{GROUPS.GOVT}}, \code{\link{Indices}},
  \code{\link{LoanMarketClimate}},
  \code{\link{LoanMarket}}, \code{\link{PHYSICAL.RISK.CB}},
  \code{\link{PHYSICAL.RISK.EQ}}, \code{\link{Receipts}},
  \code{\link{RevenueSplit}}, \code{\link{SCENLong}},
  \code{\link{SCEN}}, \code{\link{SEC.TYPE.BONDS}},
  \code{\link{SectorBridge}}, \code{\link{TYPE.BONDS}},
  \code{\link{TYPE.EQUITY}}, \code{\link{TYPE.OTHERS}},
  \code{\link{TYPE.RECEIPTS}}, \code{\link{sector.bridge}}
}
\concept{exported datasets}
\concept{possible_snapshots}
\keyword{datasets}
