% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loanbook_description.R
\name{loanbook_description}
\alias{loanbook_description}
\title{Description of a generic loanbook dataset}
\description{
Data on loans and borrowers. The information is usually owned by banks and
private.
}
\examples{
library(dplyr)

data_dictionary \%>\%
  filter(dataset == "fake_loanbook") \%>\%
  select(-1)
}
\seealso{
data_dictionary

Other bank datasets: \code{\link{fake_loanbook}}
}
\concept{bank datasets}
