% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_snapshot.R
\name{path_snapshot}
\alias{path_snapshot}
\title{Get path to data snapshots}
\usage{
path_snapshot(path = NULL)
}
\arguments{
\item{path}{The file name of a data snapshot. \code{NULL} returns the name of all
snapshots.}
}
\description{
By default, the data snapshots you take are stored in the
installation-directory of this package (inst/extdata). This function creates
paths to access them easily.
}
\examples{
path_snapshot()

this_dataset <- path_snapshot("INDEX.REGIONS.csv")
this_dataset

vroom::vroom(this_dataset)
}
\seealso{
Other handle snapshots: \code{\link{possible_snapshots}},
  \code{\link{take_snapshot}}
}
\concept{handle snapshots}
