% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/take_snapshot.R
\name{possible_snapshots}
\alias{possible_snapshots}
\title{Datasets to snapshot}
\usage{
possible_snapshots()
}
\value{
A character string giving the datasets to snapshot.
}
\description{
Datasets to snapshot
}
\examples{
possible_snapshots()
}
\seealso{
Other handle snapshots: \code{\link{path_snapshot}},
  \code{\link{take_snapshot}}
}
\concept{handle snapshots}
