% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2dii-package.R
\docType{package}
\name{r2dii-package}
\alias{r2dii}
\alias{r2dii-package}
\title{r2dii: Show 'r2dii' as a Prototype for 'tilt'}
\description{
Show 'r2dii' as a prototype for 'tilt'.
}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID})

}
\keyword{internal}
