% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2dii.dataraw-package.R
\docType{package}
\name{r2dii.dataraw-package}
\alias{r2dii.dataraw}
\alias{r2dii.dataraw-package}
\title{r2dii.dataraw: Handle Raw Or Lightly-Prepared Data}
\description{
The goal of this package is to handle raw or
    lightly-prepared data. It allows you to access "live" data from 2dii's
    dropbox folder and to take a snapshot that you can track with git.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/2DegreesInvesting/r2dii.dataraw}
  \item Report bugs at \url{https://github.com/2DegreesInvesting/r2dii.dataraw/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{mauro@2degrees-investing.org} (0000-0002-1986-7988)

Authors:
\itemize{
  \item Clare Murray \email{clare@2degrees-investing.org}
  \item Klaus Hogedorn \email{klaus@2degrees-investing.org}
}

Other contributors:
\itemize{
  \item 2 Degrees Investing Initiative \email{contact@2degrees-investing.org} [copyright holder, funder]
}

}
\keyword{internal}
