% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_version.R
\name{set_version}
\alias{set_version}
\title{Set version number as \code{major.minor.patch[.dev].datastore.snapshot}}
\usage{
set_version()
}
\value{
Modifies the DESCRIPTION file.
}
\description{
To a package version, this function appends the date when the data was
exported by datastore, then the date when the current date.
}
\examples{
# Use only interactively
# set_version()
}
\seealso{
Other miscellaneous utility functions: \code{\link{browse_private_site}},
  \code{\link{build_private_site}}, \code{\link{degrees}}
}
\concept{miscellaneous utility functions}
