% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/take_snapshot.R
\name{take_snapshot}
\alias{take_snapshot}
\title{Snapshots datasets and the configuration file}
\usage{
take_snapshot(datasets = NULL, destdir = NULL, overwrite = FALSE,
  config = get_config())
}
\arguments{
\item{datasets}{Character vector giving the name of a file (without
extension) in which to write the snapshot. It must exactly match the name
of an exported dataset or of a functions which output is a datasets. \code{NULL}
defaults to take a snapshot of all possible datasets.}

\item{destdir}{Character vector giving the path to the directory to write to.
\code{NULL} defaults to "r2dii.dataraw/inst/extdata".}

\item{overwrite}{\code{TRUE} overwrites existing snapshots (data and configuration
file).}

\item{config}{Path to a configuration file.}
}
\value{
Returns \code{NULL}. It's called for its side effect.
}
\description{
Snapshots datasets and the configuration file
}
\examples{
# Use `config` to provide a custom configuration file to `take_snapshot()`
# locally

custom_config <- example_config("config-toy.yml")
custom_config

custom_config \%>\%
  show_config()

"DebtMarketClimate" \%>\%
  take_snapshot(destdir = tempdir(), overwrite = TRUE, config = custom_config)
get_config()

# Use `options(r2dii.dataraw_config = <custom_config>)` to provide a custom
# configuration file to your R session globally.
restore <- options(r2dii.dataraw_config = custom_config)
"DebtMarketClimate" \%>\%
  take_snapshot(destdir = tempdir(), overwrite = TRUE)
get_config()

options(restore)
get_config()
}
\seealso{
\link{path_snapshot}, \link{get_config}.

Other handle snapshots: \code{\link{path_snapshot}},
  \code{\link{possible_snapshots}}
}
\concept{handle snapshots}
