% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported-data.R
\docType{data}
\name{TYPE.EQUITY}
\alias{TYPE.EQUITY}
\title{Dataset that should probably be removed or outsourced to data store}
\format{An object of class \code{character} of length 5.}
\usage{
TYPE.EQUITY
}
\description{
Dataset that should probably be removed or outsourced to data store
}
\examples{
TYPE.EQUITY
}
\seealso{
Other exported datasets to remove or outsource to data store: \code{\link{BicsSectorBridge}},
  \code{\link{GROUPS.GOVT}}, \code{\link{RevenueSplit}},
  \code{\link{SectorBridge}}, \code{\link{TYPE.OTHERS}},
  \code{\link{sector.bridge}}

Other datasets that stand alone: \code{\link{GROUPS.GOVT}},
  \code{\link{TYPE.OTHERS}}

Other possible_snapshots: \code{\link{ALD.BV}},
  \code{\link{ALD.CB}}, \code{\link{ALD.CC}},
  \code{\link{ALD.Company}}, \code{\link{ALD.EQ}},
  \code{\link{ALD.SPV}}, \code{\link{BALANCE.SHEET.DATA}},
  \code{\link{BENCH.REGIONS}},
  \code{\link{BicsSectorBridge}}, \code{\link{CB_OG}},
  \code{\link{DebtMarketClimate}},
  \code{\link{DebtMarket}}, \code{\link{EQMarket.Size}},
  \code{\link{EQ_OG}}, \code{\link{FIN.DATA}},
  \code{\link{Fund.Data}}, \code{\link{FundsTrusts}},
  \code{\link{GROUPS.GOVT}}, \code{\link{INDEX.REGIONS}},
  \code{\link{Indices}}, \code{\link{LoanMarketClimate}},
  \code{\link{LoanMarket}}, \code{\link{PHYSICAL.RISK.CB}},
  \code{\link{PHYSICAL.RISK.EQ}}, \code{\link{Receipts}},
  \code{\link{RevenueSplit}}, \code{\link{SCENLong}},
  \code{\link{SCEN}}, \code{\link{SEC.TYPE.BONDS}},
  \code{\link{SectorBridge}}, \code{\link{TYPE.BONDS}},
  \code{\link{TYPE.OTHERS}}, \code{\link{TYPE.RECEIPTS}},
  \code{\link{sector.bridge}}
}
\concept{datasets that stand alone}
\concept{exported datasets to remove or outsource to data store}
\concept{possible_snapshots}
\keyword{datasets}
